+++ fiddles/angular2/fiddle-0051-KunovskyGridColGroups/app/meta.ts
@@ -1,0 +1,7 @@
+var metadata = {
+     fiddleHeader: 'Angular 2 - fiddle-0048-SlickGridTree',
+     urls: {
+          github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0051-KunovskyGridColGroups'
+     },
+     consoleTag: 'H O U S E  ~  f i d d l e s'
+};