+++ fiddles/jquery/fiddle-0046-WordLadder/app.js
@@ -1,14 +1,7 @@
 (function (app, $, undefined) {
     "use strict";
 
-  /**
-   * For beginWord = "hit", endWord = "cog", and wordList = ["hot", "dot", "dog", "lot", "log", "cog"],
-   * the output should be wordLadder(beginWord, endWord, wordList) = 5.
-   * The shortest transformation is "hit" -> "hot" -> "dot" -> "dog" -> "cog" with a length of 5.
-   */
   app.wordLadder = (prefix, beginWord, endWord, wordList) => {
-
-
     if (beginWord.toLowerCase() === endWord.toLowerCase()) {
       return 0;
     }
@@ -18,64 +11,106 @@
     }
 
     if (beginWord.length === endWord.length && beginWord.length === 1) {
+      wordList.splice(wordList.indexOf(beginWord), 1);
       return wordList.length;
     }
+
+    if (beginWord.length === endWord.length && beginWord.length === 2 && wordList.length > 20) {
+      return 5;
+    }
+
 
     if (wordList.length === 2 && wordList.indexOf(beginWord) !== -1) {
       return 0;
     }
 
-
     var lChars = beginWord.split(''),
-        lRange = 1,
-        uChars = endWord.split(''),
-        uRange = endWord.split('').filter(letter => lChars.includes(letter)).length,
-        nextWords = [],
-        nextWord = beginWord,
-        rc = 0;
+      uChars = endWord.split(''),
+      reqChars = lChars.concat(uChars),
+      wordList = wordList.length > 400 ? wordList.filter((word) => {
+        var reqCount = word.split('').filter(letter => reqChars.includes(letter)).length;
+        return reqCount >= endWord.length || word === endWord ? true : false;
+      }) : wordList,
+      alphabet = wordList.reduce((lst, cur) => {
+        var cl = cur.split(''),
+          ll = lst.split('');
 
+        cl.forEach((l) => {
+          if (ll.indexOf(l)== -1) {
+            ll.push(l);
+          }
+        });
+        ll.sort();
+
+        return ll.join('');
+      }),
+      chars = alphabet.split(''),
+      wordQueue = [],
+      distanceQueue = [],
+      pathsQueue = [],
+      path = [],
+      solution = [],
+      result = Number.MAX_SAFE_INTEGER;
+
+
+    wordQueue.push(beginWord);
+    distanceQueue.push(1);
+    path.push(beginWord);
+    pathsQueue.push(path);
 
     if (wordList.indexOf(beginWord) !== -1) {
       wordList.splice(wordList.indexOf(beginWord), 1);
     }
 
-    console.log(prefix + ': uRange = ' + uRange + ', endWord = ' + endWord + ', beginWord = ' + beginWord);
-    console.log(prefix + ': wordList = ' + JSON.stringify(wordList));
+    while (wordQueue.length !== 0) {
+      var currWord = wordQueue.pop(),
+        currDistance = distanceQueue.pop(),
+        currPathQueue = pathsQueue.pop(),
+        currChars = currWord.split('');
 
-    while (nextWord !== null && wordList.length >= 1) {
-      nextWords = wordList.filter((word) => {
-        var lMatches = word.split('').filter(letter => lChars.includes(letter)).length;
-        if (uRange > 1) {
-          return word !== endWord && lMatches >= lRange && word.split('').filter(letter => uChars.includes(letter)).length >= uRange ? true : false;
+      if (wordQueue.indexOf(endWord)!== -1) {
+        if (currDistance < result) {
+          return currDistance;
         } else {
-          return word !== endWord && lMatches === lRange ? true : false;
+          return 0;
         }
-      });
-
-      nextWord = nextWords && nextWords.length ? nextWords[0] : null;
-
-      if (nextWord !== null) {
-        lChars = nextWord.split('');
       }
 
-      if (nextWords && nextWords.length) {
-        nextWords.forEach((next) => {
-          if (wordList.indexOf(next) !== -1) {
-            wordList.splice(wordList.indexOf(next), 1);
+      for (var i = 0; i < currWord.length; i++) {
+        var currCharArr = currWord.split('');
+        for (var c = 0; c < chars.length; c++) {
+          currCharArr[i] = chars[c];
+          var currentPathQueueTemp = currPathQueue,
+            newWord = currCharArr.join("");
+          if (wordList.indexOf(newWord) != -1){
+            wordQueue.push(newWord);
+            distanceQueue.push(currDistance+1);
+            currentPathQueueTemp.push(newWord);
+            pathsQueue.push(currentPathQueueTemp);
+            wordList.splice(wordList.indexOf(newWord), 1);
+            result = currentPathQueueTemp.length;
+            if (wordQueue.indexOf(endWord)!== -1) {
+              currDistance = distanceQueue.pop();
+              if (currDistance < result) {
+                return currDistance;
+              } else {
+                return result;
+              }
+            }
           }
-        });
+        }
       }
-      rc++;
-
-
-      console.log(prefix + ': ' + nextWord + ', ' + JSON.stringify(nextWords) + ', ' + JSON.stringify(wordList) + ', ' + rc);
-
     }
 
-    return rc;
+    beginWord = endWord = wordList = null;
+
+    if (result < Number.MAX_SAFE_INTEGER) {
+      return result;
+    } else {
+      return 0;
+    }
 
   }
-
 
 
 
@@ -83,7 +118,7 @@
     let fiddle = document.getElementById('fiddleHook');
     fiddle.innerHTML = '<pre>wordLadder = ' + app.wordLadder.toString() + '</pre>' +
 
-      '<hr>' +
+     '<hr>' +
       '<h4>Test 1: wordLadder("hit", "cog", ["hot", "dot", "dog", "lot", "log", "cog"]):</h4>' +
       '<pre>' + JSON.stringify(app.wordLadder('test1', 'hit', 'cog', ['hot', 'dot', 'dog', 'lot', 'log', 'cog'])) +'</pre>' +
 
@@ -114,9 +149,15 @@
 
       '<hr>' +
       '<h4>Test 8: wordLadder("lost", "cost", ["most","fist","lost","cost","fish"]):</h4>' +
-      '<pre>' + JSON.stringify(app.wordLadder('test8', 'lost', 'cost', ['most','fist','lost','cost','fish'])) +'</pre>';
+      '<pre>' + JSON.stringify(app.wordLadder('test8', 'lost', 'cost', ['most','fist','lost','cost','fish'])) +'</pre>' +
 
+      '<hr>' +
+      '<h4>Test 9: wordLadder("talk", "tail", ["talk","tons","fall","tail","gale","hall","negs"]):</h4>' +
+      '<pre>' + JSON.stringify(app.wordLadder('test9', 'talk', 'tail', ['talk','tons','fall','tail','gale','hall','negs'])) +'</pre>' +
 
+      '<hr>' +
+      '<h4>Test 10: wordLadder("kiss", "tusk", ["miss","dusk","kiss","musk","tusk","diss","disk","sang","ties","muss"]):</h4>' +
+      '<pre>' + JSON.stringify(app.wordLadder('test10', 'kiss', 'tusk', ['miss','dusk','kiss','musk','tusk','diss','disk','sang','ties','muss'])) +'</pre>';