+++ fiddles/jquery/fiddle-0048-AddTwoHugeNumbers/app.js
@@ -2,84 +2,129 @@
     "use strict";
 
     app.addTwoHugeNumbers = (prefix, a, b) => {
-    "use strict";
+      "use strict";
 
-    let x = JSON.stringify(a).slice(1,-1).split(','),
-      y = JSON.stringify(b).slice(1,-1).split(','),
-      u = '',
-      v = '',
-      rc = [],
-      sum = '',
-      uSeg = 0,
-      lSeg = 0,
-      iter = 0,
-      carry = 0,
-      i = 0;
+      let x = JSON.stringify(a).slice(1,-1).split(','),
+        y = JSON.stringify(b).slice(1,-1).split(','),
+        u = '',
+        v = '',
+        rc = [],
+        sum = '',
+        uSeg = 0,
+        lSeg = 0,
+        iter = 0,
+        carry = 0,
+        i = 0,
+        buffer = {
+          a: 0,
+          b: 0,
+          sum: 0,
+          rem: 0
+        },
+        calc = function(buf) {
+          var u = (+buf.a),
+          v = (+buf.b),
+          z = (+buf.rem),
+          sum = u + v + z;
 
-    if (typeof a === 'object' && typeof b === 'object') {
-      if // same size
-      (x.length === y.length && x.length > 1) {
-        y.forEach((digit, index) => {
-          rc.push((+digit) + (+x[index]));
-      });
-      } // both zero
-      else if (y.length === x.length && x.length === 1 && x[0] === y[0] && x[0] === "0") {
-        rc = x;
-      } // a is zero
-      else if(y.length > 1 && x.length === 1 && x[0] === "0") {
-        rc = y;
-      } // b is zero
-      else if(x.length > 1 && y.length === 1 && y[0] === "0") {
-        rc = x;
-      } // otherwise combine, add, divide
-      else {
-        u = JSON.stringify(a).slice(1,-1).replace(/,/g,'');
-        v = JSON.stringify(b).slice(1,-1).replace(/,/g,'');
-        sum = Number((+u) + (+v)) + "";
+          var sumStr = String(sum),
+            carryLeftStr = sumStr.length > 4 ? sumStr.substr(0, sumStr.length - 4) : null,
+            right4DigitsStr = sumStr.length > 4 ? sumStr.substr(sumStr.length - 4, sumStr.length) : sumStr,
+            carryLeft = carryLeftStr ? +carryLeftStr : null;
 
-        if // sum is exponential
-        (sum.indexOf('e') !== -1) {
-          let left = sum.substring(0, sum.indexOf('e')),
-            right = sum.substring(sum.indexOf('+') + 1, sum.length),
-            hasDecimal = left.indexOf('.') !== -1 ? true : false,
-            decimalOffset = hasDecimal ? left.length - (left.indexOf('.') + 1) : 0,
-            zeroCount = +right - decimalOffset;
+          buf.sum = right4DigitsStr;
 
-          if (hasDecimal) {
-            left = left.replace('.', '');
+          if (carryLeft) {
+            buf.rem = carryLeftStr;
+          } else {
+            buf.rem = '0';
           }
 
-          sum = left;
+          return buf;
+        },
+        buf = Object.assign(buffer, {});
 
-          while(i < zeroCount) {
-            sum += '0';
-            i++;
+
+      if (typeof a === 'object' && typeof b === 'object') {
+        if // same size
+        (x.length === y.length && x.length > 1) {
+          rc = y;
+          i = y.length - 1;
+          for(var xi = x.length -1; xi >= 0; xi--) {
+            buf.a = y[i];
+            buf.b = x[xi];
+            buf = calc(buf);
+            rc[i] = buf.sum;
+            i--;
           }
-          i = 0;
+          if (buf.rem !== '0') {
+            rc = [buf.rem].concat(rc);
+          }
+        } // both zero
+        else if (y.length === x.length && x.length === 1 && x[0] === y[0] && x[0] === "0") {
+          rc = x;
+        } // a is zero
+        else if(y.length > 1 && x.length === 1 && x[0] === "0") {
+          rc = y;
+        } // b is zero
+        else if(x.length > 1 && y.length === 1 && y[0] === "0") {
+          rc = x;
+        } // x is smaller than y
+        else if (x.length < y.length) {
+            rc = y;
+            i = y.length - 1;
+            for(var xi = x.length -1; xi >= 0; xi--) {
+              buf.a = y[i];
+              buf.b = x[xi];
+              buf = calc(buf);
+              rc[i] = buf.sum;
+              i--;
+            }
 
-        } // end if sum is ...
+            while (i >= 0 && buf.rem !== '0') {
+              buf.a = y[i];
+              buf.b = buf.rem;
+              buf.rem = '0';
+              buf = calc(buf);
+              rc[i] = buf.sum;
+              i--;
+            }
+            if (buf.rem !== '0') {
+              rc = [buf.rem].concat(rc);
+            }
 
+        } // y is smaller than x
+        else {
+          rc = x;
+          i = x.length - 1;
+          for(var yi = y.length -1; yi >= 0; yi--) {
+            buf.a = x[i];
+            buf.b = y[yi];
+            buf = calc(buf);
+            rc[i] = buf.sum;
+            i--;
+          }
 
-        carry = sum.length % 4;
-        iter = sum.length / 4;
-        lSeg = sum.length - 4;
-        uSeg = sum.length;
+          while (i >= 0 && buf.rem !== '0') {
+            buf.a = x[i];
+            buf.b = buf.rem;
+            buf.rem = '0';
+            buf = calc(buf);
+            rc[i] = buf.sum;
+            i--;
+          }
 
-        while (i < iter) {
-          rc.unshift(sum.substring(lSeg, uSeg));
-          lSeg -= 4;
-          uSeg -= 4;
-          i++;
-        } // while
+          if (buf.rem !== '0') {
+            rc = [buf.rem].concat(rc);
+          }
 
-      } // end else
+        }
 
-      a = b = null;
+        a = b = null;
 
-      rc.map(Number);
-    }
+      }
 
-    return rc;
+      return rc.map(Number);
 
     } // end func
 
@@ -118,6 +163,8 @@
     '<hr>' +
     '<h4>Test 8: addTwoHugeNumbers([8339, 4510],  [2309]):</h4>' +
     '<pre>' + JSON.stringify(app.addTwoHugeNumbers('test 8', [8339, 4510],  [2309])) +'</pre>';
+
+
 
   });
 })(window.app = window.app || {}, jQuery)